+++ tests/AbstractTestCase.php
@@ -20,6 +20,11 @@ use PHPUnit\Framework\TestCase;
 
 abstract class AbstractTestCase extends TestCase
 {
+    protected function setUp(): void
+    {
+        $_SERVER['HTTP_X_FORWARDED_PREFIX'] = null;
+    }
+
     /**
      * Get app.
      *