+++ tests/Helper/RouterHelperTraitTest.php
@@ -13,16 +13,16 @@
 namespace Berlioz\Http\Core\Tests\Helper;
 
 use Berlioz\Core\Core;
-use Berlioz\Http\Message\ServerRequest;
 use Berlioz\Http\Core\App\HttpApp;
 use Berlioz\Http\Core\Helper\RouterHelperTrait;
 use Berlioz\Http\Core\TestProject\FakeDefaultDirectories;
+use Berlioz\Http\Core\Tests\AbstractTestCase;
+use Berlioz\Http\Message\ServerRequest;
 use Berlioz\Router\Exception\RoutingException;
 use Berlioz\Router\Route;
 use Berlioz\Router\RouterInterface;
-use PHPUnit\Framework\TestCase;
 
-class RouterHelperTraitTest extends TestCase
+class RouterHelperTraitTest extends AbstractTestCase
 {
     private function getHelper()
     {
@@ -67,6 +67,23 @@ class RouterHelperTraitTest extends TestCase
         $helper = $this->getHelper();
 
         $this->assertEquals('/controller1/method1', (string)$helper->path('c1m1'));
+    }
+
+    public function testPath_withPrefix()
+    {
+        $_SERVER['HTTP_X_FORWARDED_PREFIX'] = '/super-prefix';
+
+        $helper = $this->getHelper();
+
+        $this->assertEquals('/super-prefix/controller1/method1', (string)$helper->path('c1m1'));
+    }
+
+    public function testPath_withRoute()
+    {
+        $helper = $this->getHelper();
+
+        $router = $helper->getApp()->getRouter();
+        $this->assertEquals('/controller1/method1', (string)$helper->path($router->getRoute('c1m1')));
     }
 
     public function testPath_missingAttribute()