+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file. This projec
 to [Semantic Versioning] (http://semver.org/). For change log format,
 use [Keep a Changelog] (http://keepachangelog.com/).
 
+## [2.3.0] - 2024-11-04
+
+### Changed
+
+- Bump library `berlioz/core` to ^2.4
+- `RouterHelperTrait::path()` accepts `RouteInterface` or string instead of only a string
+- New options: "Berlioz.router" to give parameters to the Router
+
 ## [2.2.7] - 2024-10-04
 
 ### Fixed